[REFACTOR] use hooks for functionaltity
